+++ extensions/ql-vscode/src/common/invocation-rate-limiter.ts
@@ -1,4 +1,4 @@
-import { Memento } from "./common/memento";
+import { Memento } from "./memento";
 
 /**
  * Provides a utility method to invoke a function only if a minimum time interval has elapsed since
